radlab-ui/automation/terraform/infrastructure/function/email_notification/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
functions-framework==3.*
tabulate